requirements.txt
python-dotenv==1.0.1
qdrant-client==1.12.1
sentence-transformers==3.2.1
pymupdf==1.24.14
python-docx==1.1.2
openpyxl>=3.1.2
pandas>=2.0.0
numpy>=1.26.0
rank-bm25>=0.2.2
openai>=1.0.0
requests==2.32.3
pytest==8.3.5
